Added bypass of trigger check if fault is instruction fault and trigger hitcounter is zero
In this corner case the tool will directly inject the fault at start when the first tb is analysed.
If the condition is met the trigger address is a dont care and can be ignored.
Therefore it is a special type and is an exception to the check. The corresponding function is called handle_first_tb_fault_insertion and can be found in faultpluing/faultplugin.c
Fixes issue #18