Fix handling of the trigger hitcounter
The plugin and python code currently have different understandings of
what the hitcounter represents. A value of 1 on the plugin side means
that a fault is injected with the first execution of the trigger
address. The python side handles a 1 as if the execution after the first
execution of the trigger address (the second overall) should be faulted.

Fix this by handling the hitcounter in the python in the same way as in
the plugin. Also adjust the fault readme to clarify this.