Fix trigger calculation if it is in a diffrent TB than the fault address
If the trigger address is in a different TB than the fault address, the
fault is triggered incorrectly in some situations. This occurs, if the
TB with the trigger address is executed independently of the TB with the
fault location. The hitcounter of the trigger will not match the point
at which the fault address should be faulted in this case.

Adjust the hitcounter of the trigger so it matches the correct fault
location specified in the fault configuration.