Make analysis flake8 compliant (#36)
* Reformat file to make it black compatible

* Added sections that are not reformated by black

This is to preserve the matrix representation. Black otherwise would simply make a long vector which is hard to read

* Remove flake8 E741 issue

Renamed l to log to remove E741 problem. On some editors l could be be mistaken for I.

* Run black on *.py files in analysis

Save changes of black in repository

* Update lint worker

Update lint worker to check analysis folder