Specify explicit register data type to avoid integer conversion errors
In situations where no data type is explicitly specified, pandas may
choose to convert integers close to the maximum 64-bit integer value to
floats during to_dict() calls. This is a common occurrence for the
register dataset of 64-bit architectures and will lead to type errors.

Explicitly specify registers to be of type "UInt64" to avoid integer
conversion.