Fix goldenrun when an end point other than the first is hit
The goldenrun is currently determined to be successful if the first end
point is hit. Otherwise, it is assumed that the maximum number of TBs
was executed without hitting an end point. With support for multiple end
points this is not accurate anymore. Instead use the end_reason
attribute to determine if the goldenrun has hit the max TB timeout.